requirements.txt
django
requests
beautifulsoup4
selenium
django-background-tasks